Refactor out the testing functions so we can use this for benchmarking
